:recycle: Rename package name for api
