+++ api/handler_test.go
@@ -1,4 +1,4 @@
-package pipeline
+package api
 
 import (
 	"encoding/json"
